Add files via upload
This is a website for a pastry shop, designed using HTML, CSS, JavaScript, jQuery, and Bootstrap. To view the site, please follow these steps:

Download the zipped file.
Extract the contents onto your computer.
Open the extracted folder.
Double-click on index.html to launch the website.
You will be able to explore the website. Navigate through the various options and menus to access other pages.